+++ src/br/com/etyllica/core/application/GenericLoadApplication.java
@@ -47,7 +47,7 @@ public class GenericLoadApplication extends DefaultLoadApplication{
 		g.drawStringShadowX(280-y, phrase);
 
 		g.drawRect(rectX, rectY, rectW, rectH);
-		g.fillRect(rectX+2, rectY+2, ((rectW*fill)/100)-3, rectH-3);
+		g.fillRect(rectX+2, rectY+2, (int)((rectW*fill)/100)-3, rectH-3);
 		
 		g.setFont(p);
 		g.drawStringShadow(rectX, rectY, rectW, rectH, percent, Color.BLACK);
